adding documentation to cubes n triangles
