--- app/src/main/java/com/example/clicker/farmingGame/MyGLRenderer.kt
@@ -1,23 +1,0 @@
-package com.example.clicker.farmingGame
-
-import android.opengl.EGLConfig
-import android.opengl.GLES20
-import android.opengl.GLSurfaceView
-import javax.microedition.khronos.opengles.GL10
-
-class MyGLRenderer : GLSurfaceView.Renderer {
-
-
-    override fun onDrawFrame(unused: GL10) {
-        // Redraw background color
-        GLES20.glClear(GLES20.GL_COLOR_BUFFER_BIT)
-    }
-
-    override fun onSurfaceCreated(p0: GL10?, p1: javax.microedition.khronos.egl.EGLConfig?) {
-        GLES20.glClearColor(0.0f, 1.0f, 0.0f, 1.0f)
-    }
-
-    override fun onSurfaceChanged(unused: GL10, width: Int, height: Int) {
-        GLES20.glViewport(0, 0, width, height)
-    }
-}
